Moved StaticRoutes to the common web server module
